Add api key working
