+++ .gitignore
@@ -16,6 +16,8 @@
 .env.test.local
 .env.production.local
 
+apikey.js
+
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
