+++ src/TenDay/TenDay.js
@@ -5,14 +5,10 @@ import './tenday.css';
 
 function TenDay(props) {
 
-  const tenDayArray = props.data.filter(period=>{
-    return !period.title.includes('Night');
-  })
-
   if(props.toggleState) {
     return (
       <article className='ten-day'>
-        {tenDayArray.map((day) => {
+        {props.data.map((day) => {
           return <Daily  data={day}
                           key={day.period} />
         })}
