+++ src/current/current.css
@@ -1,6 +1,6 @@
 @import url('https://fonts.googleapis.com/css?family=Open+Sans:300,800');
 
-.current {
+.current-weather-data {
 	display: flex;
 	justify-content: space-around;
 	width: 83%;
@@ -10,7 +10,21 @@
   font-family: 'Open Sans', sans-serif;
   color: #C9C9C9;
   background:rgba(0,0,0,0.5);
-}e
+}
+
+.current-location {
+  font-size: 1.5rem;
+  margin: auto;
+  padding-top: 1rem;
+  text-align: center;
+  color: #C9C9C9;
+  background:rgba(0,0,0,0.5);
+  max-width: 32rem;
+}
+
+.day-label {
+  font-weight: 300;
+}
 
 article {
   margin-right: 2rem;
@@ -29,7 +43,7 @@ article {
 
 .temperature {
   font-size: 6rem;
-  margin: 0 0 1rem 0;
+  margin: 0;
   color: #FFF;
   font-weight: 300;
 }
@@ -44,8 +58,8 @@ article {
 .highlow {
   font-size: 1.2rem;
   font-weight: 300;
-  margin-top: 1.5rem;
-  margin-bottom: 1.5rem;
+  margin-top: 1rem;
+  margin-bottom: 1rem;
 }
 
 .weather-label,
